+++ syntax/vm/vm.go
@@ -12,6 +12,7 @@ import (
 	"github.com/grafana/alloy/syntax/internal/stdlib"
 	"github.com/grafana/alloy/syntax/internal/syntaxtags"
 	"github.com/grafana/alloy/syntax/internal/tagcache"
+	"github.com/grafana/alloy/syntax/internal/transform"
 	"github.com/grafana/alloy/syntax/internal/value"
 )
 
@@ -303,8 +304,7 @@ func (vm *Evaluator) evaluateExpr(scope *Scope, assoc map[value.Value]ast.Node, 
 
 	switch expr := expr.(type) {
 	case *ast.LiteralExpr:
-		return valueFromLiteral(expr.Value, expr.Kind)
-
+		return transform.ValueFromLiteral(expr.Value, expr.Kind)
 	case *ast.BinaryExpr:
 		lhs, err := vm.evaluateExpr(scope, assoc, expr.Left)
 		if err != nil {
